LayoutTests/imported/w3c/web-platform-tests/fetch/redirects/subresource-fragments-expected.txt


FAIL Control assert_approx_equals: Red channel of the pixel at (40, 40) expected 0 +/- undefined but got 0
FAIL Redirect with the original URL containing a fragment assert_approx_equals: Red channel of the pixel at (40, 40) expected 0 +/- undefined but got 0
FAIL Redirect with the response Location header containing a fragment assert_approx_equals: Red channel of the pixel at (40, 40) expected 0 +/- undefined but got 0
FAIL Redirect with both the original URL and response Location header containing a fragment assert_approx_equals: Red channel of the pixel at (40, 40) expected 0 +/- undefined but got 0
